the big fix
it all works now

+++ src/main/java/com/apw/sbcio/PWMController.java
@@ -23,7 +23,7 @@ public interface PWMController {
    * implementation.
    * @param angle Double from 0 to 180 inclusive representing the angle of the servo.
    */
-  default void setServoAngle(int pin, double angle) {
+  default void setServoAngle(int pin, int angle) {
     setOutputPulseWidth(pin, 1 + angle / 180);
   }
